+++ axelor-account/src/main/java/com/axelor/apps/account/web/InvoiceLineController.java
@@ -22,6 +22,7 @@ import java.util.Map;
 
 import com.axelor.apps.account.db.Invoice;
 import com.axelor.apps.account.db.InvoiceLine;
+import com.axelor.apps.account.db.repo.InvoiceLineRepository;
 import com.axelor.apps.account.service.app.AppAccountService;
 import com.axelor.apps.account.service.invoice.InvoiceLineService;
 import com.axelor.apps.account.service.invoice.generator.line.InvoiceLineManagement;
@@ -62,8 +63,13 @@ public class InvoiceLineController {
 	public void computeAnalyticDistribution(ActionRequest request, ActionResponse response) throws AxelorException{
 		InvoiceLine invoiceLine = request.getContext().asType(InvoiceLine.class);
 		Invoice invoice = invoiceLine.getInvoice();
+		Context context = request.getContext();
+		
 		if(invoice == null){
-			invoice = request.getContext().getParent().asType(Invoice.class);
+			if(context.getParent().getContextClass() == InvoiceLine.class) {
+				context = request.getContext().getParent();
+			}
+			invoice = context.getParent().asType(Invoice.class);
 			invoiceLine.setInvoice(invoice);
 		}
 		if(Beans.get(AppAccountService.class).getAppAccount().getManageAnalyticAccounting()){
@@ -75,9 +81,12 @@ public class InvoiceLineController {
 	public void compute(ActionRequest request, ActionResponse response) throws AxelorException {
 
 		Context context = request.getContext();
-		
 		InvoiceLine invoiceLine = context.asType(InvoiceLine.class);
-
+		
+		if(context.getParent().getContextClass() == InvoiceLine.class) {
+			context = request.getContext().getParent();
+		}
+		
 		Invoice invoice = this.getInvoice(context);
 		
 		if(invoice == null || invoiceLine.getPrice() == null || invoiceLine.getQty() == null)  {  return;  }
@@ -120,6 +129,11 @@ public class InvoiceLineController {
 	public void getProductInformation(ActionRequest request, ActionResponse response) throws AxelorException {
 		Context context = request.getContext();
 		InvoiceLine invoiceLine = context.asType(InvoiceLine.class);
+		
+		if(context.getParent().getContextClass() == InvoiceLine.class) {
+			context = request.getContext().getParent();
+		}
+
 		Invoice invoice = this.getInvoice(context);
 		Product product = invoiceLine.getProduct();
 		Map<String, Object> productInformation = invoiceLineService.resetProductInformation();
@@ -138,8 +152,10 @@ public class InvoiceLineController {
 	public void getDiscount(ActionRequest request, ActionResponse response) throws AxelorException {
 
 		Context context = request.getContext();
-		
 		InvoiceLine invoiceLine = context.asType(InvoiceLine.class);
+		if(context.getParent().getContextClass() == InvoiceLine.class) {
+			context = request.getContext().getParent();
+		}
 
 		Invoice invoice = this.getInvoice(context);
 		
@@ -167,8 +183,10 @@ public class InvoiceLineController {
 	public void convertUnitPrice(ActionRequest request, ActionResponse response) {
 
 		Context context = request.getContext();
-		
 		InvoiceLine invoiceLine = context.asType(InvoiceLine.class);
+		if(context.getParent().getContextClass() == InvoiceLine.class) {
+			context = request.getContext().getParent();
+		}
 
 		Invoice invoice = this.getInvoice(context);
 		
@@ -199,16 +217,15 @@ public class InvoiceLineController {
 	
 	public void emptyLine(ActionRequest request, ActionResponse response){
 		InvoiceLine invoiceLine = request.getContext().asType(InvoiceLine.class);
-		if(invoiceLine.getIsTitleLine()){
-			InvoiceLine newInvoiceLine = new InvoiceLine();
-			newInvoiceLine.setIsTitleLine(true);
-			newInvoiceLine.setQty(BigDecimal.ZERO);
-			newInvoiceLine.setId(invoiceLine.getId());
-			newInvoiceLine.setVersion(invoiceLine.getVersion());
-			response.setValues(Mapper.toMap(newInvoiceLine));
+		if(invoiceLine.getTypeSelect() != InvoiceLineRepository.TYPE_NORMAL){
+			Map<String,Object> newInvoiceLine =  Mapper.toMap(new InvoiceLine());
+			newInvoiceLine.put("qty", BigDecimal.ZERO);
+			newInvoiceLine.put("id", invoiceLine.getId());
+			newInvoiceLine.put("version", invoiceLine.getVersion());
+			newInvoiceLine.put("typeSelect", invoiceLine.getTypeSelect());
+			response.setValues(newInvoiceLine);
 		}
 	}
-	
 	
 	public Invoice getInvoice(Context context)  {
 		
@@ -225,5 +242,4 @@ public class InvoiceLineController {
 		
 		return invoice;
 	}
-
 }