+++ axelor-business-project/src/main/java/com/axelor/apps/businessproject/service/InvoicingProjectService.java
@@ -175,7 +175,7 @@ public class InvoicingProjectService {
 
 		InvoiceLineGenerator invoiceLineGenerator = new InvoiceLineGeneratorSupplyChain(invoice, product, saleOrderLine.getProductName(),
 				saleOrderLine.getDescription(), saleOrderLine.getQty(), saleOrderLine.getUnit(),
-				priority, false, saleOrderLine, null, null)  {
+				priority, false, saleOrderLine, null, null, false)  {
 
 			@Override
 			public List<InvoiceLine> creates() throws AxelorException {
@@ -210,7 +210,7 @@ public class InvoicingProjectService {
 
 		InvoiceLineGeneratorSupplyChain invoiceLineGenerator = new InvoiceLineGeneratorSupplyChain(invoice, product, purchaseOrderLine.getProductName(),
 				purchaseOrderLine.getDescription(), purchaseOrderLine.getQty(), purchaseOrderLine.getUnit(),
-				priority, false, null, purchaseOrderLine, null)  {
+				priority, false, null, purchaseOrderLine, null, false)  {
 			@Override
 			public List<InvoiceLine> creates() throws AxelorException {
 
@@ -252,7 +252,7 @@ public class InvoicingProjectService {
 
 		InvoiceLineGenerator invoiceLineGenerator = new InvoiceLineGenerator(invoice, product, project.getName(), project.getPrice(),
 					project.getPrice(), null, project.getQty(), project.getUnit(), null, priority, BigDecimal.ZERO, PriceListLineRepository.AMOUNT_TYPE_NONE,
-					project.getPrice().multiply(project.getQty()), null,false)  {
+					project.getPrice().multiply(project.getQty()), null,false, false)  {
 
 			@Override
 			public List<InvoiceLine> creates() throws AxelorException {
@@ -308,36 +308,38 @@ public class InvoicingProjectService {
 		return;
 	}
 	
+	
+	//TODO Replace hard coded value per static variable
 	public void fillLines(InvoicingProject invoicingProject,Project project){
 		if(project.getProjInvTypeSelect() == ProjectRepository.INVOICING_TYPE_FLAT_RATE || project.getProjInvTypeSelect() == ProjectRepository.INVOICING_TYPE_TIME_BASED)  {
 			if (invoicingProject.getDeadlineDate() != null){
 				invoicingProject.getSaleOrderLineSet().addAll(Beans.get(SaleOrderLineRepository.class)
-						.all().filter("self.project = ?1 AND self.toInvoice = true AND self.invoiced = false AND self.saleOrder.creationDate < ?2", project, invoicingProject.getDeadlineDate()).fetch());
+						.all().filter("self.project = ?1 AND self.toInvoice = true AND self.invoiced = false AND self.saleOrder.creationDate < ?2 AND (self.saleOrder.statusSelect = 3 OR self.saleOrder.statusSelect = 4)", project, invoicingProject.getDeadlineDate()).fetch());
 				
 				invoicingProject.getPurchaseOrderLineSet().addAll(Beans.get(PurchaseOrderLineRepository.class)
-						.all().filter("self.project = ?1 AND self.toInvoice = true AND self.invoiced = false AND self.purchaseOrder.orderDate < ?2", project, invoicingProject.getDeadlineDate()).fetch());
+						.all().filter("self.project = ?1 AND self.toInvoice = true AND self.invoiced = false AND self.purchaseOrder.orderDate < ?2 AND (self.purchaseOrder.statusSelect = 3 OR self.purchaseOrder.statusSelect = 4)", project, invoicingProject.getDeadlineDate()).fetch());
 				
 				invoicingProject.getLogTimesSet().addAll(Beans.get(TimesheetLineRepository.class)
 						.all().filter("self.timesheet.statusSelect = 3 AND self.project = ?1 AND self.toInvoice = true AND self.invoiced = false AND self.date < ?2", project,invoicingProject.getDeadlineDate()).fetch());
 				
 				invoicingProject.getExpenseLineSet().addAll(Beans.get(ExpenseLineRepository.class)
-						.all().filter("self.project = ?1 AND self.toInvoice = true AND self.invoiced = false AND self.expenseDate < ?2", project, invoicingProject.getDeadlineDate()).fetch());
+						.all().filter("self.project = ?1 AND self.toInvoice = true AND self.invoiced = false AND self.expenseDate < ?2 AND (self.expense.statusSelect = 3 OR self.expense.statusSelect = 4)", project, invoicingProject.getDeadlineDate()).fetch());
 				
 				invoicingProject.getElementsToInvoiceSet().addAll(Beans.get(ElementsToInvoiceRepository.class)
 						.all().filter("self.project = ?1 AND self.toInvoice = true AND self.invoiced = false AND self.date < ?2", project,invoicingProject.getDeadlineDate()).fetch());
 			}
 			else {
 				invoicingProject.getSaleOrderLineSet().addAll(Beans.get(SaleOrderLineRepository.class)
-						.all().filter("self.project = ?1 AND self.toInvoice = true AND self.invoiced = false", project).fetch());
+						.all().filter("self.project = ?1 AND self.toInvoice = true AND self.invoiced = false AND (self.saleOrder.statusSelect = 3 OR self.saleOrder.statusSelect = 4)", project).fetch());
 				
 				invoicingProject.getPurchaseOrderLineSet().addAll(Beans.get(PurchaseOrderLineRepository.class)
-						.all().filter("self.project = ?1 AND self.toInvoice = true AND self.invoiced = false", project).fetch());
+						.all().filter("self.project = ?1 AND self.toInvoice = true AND self.invoiced = false AND (self.purchaseOrder.statusSelect = 3 OR self.purchaseOrder.statusSelect = 4)", project).fetch());
 				
 				invoicingProject.getLogTimesSet().addAll(Beans.get(TimesheetLineRepository.class)
 						.all().filter("self.timesheet.statusSelect = 3 AND self.project = ?1 AND self.toInvoice = true AND self.invoiced = false", project).fetch());
 				
 				invoicingProject.getExpenseLineSet().addAll(Beans.get(ExpenseLineRepository.class)
-						.all().filter("self.project = ?1 AND self.toInvoice = true AND self.invoiced = false", project).fetch());
+						.all().filter("self.project = ?1 AND self.toInvoice = true AND self.invoiced = false AND (self.expense.statusSelect = 3 OR self.expense.statusSelect = 4)", project).fetch());
 				
 				invoicingProject.getElementsToInvoiceSet().addAll(Beans.get(ElementsToInvoiceRepository.class)
 						.all().filter("self.project = ?1 AND self.toInvoice = true AND self.invoiced = false", project).fetch());