+++ axelor-sale/src/main/java/com/axelor/apps/sale/service/saleorder/SaleOrderComputeService.java
@@ -18,8 +18,10 @@
 package com.axelor.apps.sale.service.saleorder;
 
 import java.math.BigDecimal;
+import java.util.List;
 
 import com.axelor.apps.sale.db.SaleOrder;
+import com.axelor.apps.sale.db.SaleOrderLine;
 import com.axelor.exception.AxelorException;
 
 public interface SaleOrderComputeService {
@@ -72,5 +74,8 @@ public interface SaleOrderComputeService {
 	 * @return  total price from the sale order lines
 	 */
 	public BigDecimal getTotalSaleOrderPrice(SaleOrder saleOrder);
+
+	
+  public List<SaleOrderLine> removeSubLines(List<SaleOrderLine> lines);
 
 }