+++ axelor-stock/src/main/java/com/axelor/apps/stock/web/StockMoveLineController.java
@@ -17,8 +17,11 @@
  */
 package com.axelor.apps.stock.web;
 
+import java.math.BigDecimal;
+import java.util.Map;
 import com.axelor.apps.stock.db.StockMove;
 import com.axelor.apps.stock.db.StockMoveLine;
+import com.axelor.apps.stock.db.repo.StockMoveLineRepository;
 import com.axelor.apps.stock.service.StockMoveLineService;
 import com.axelor.db.mapper.Mapper;
 import com.axelor.exception.AxelorException;
@@ -39,14 +42,20 @@ public class StockMoveLineController {
 	public void compute(ActionRequest request, ActionResponse response) throws AxelorException {
 		StockMoveLine stockMoveLine = request.getContext().asType(StockMoveLine.class);
 		StockMove stockMove = stockMoveLine.getStockMove();
+		
 		if(stockMove == null){
 			Context parentContext = request.getContext().getParent();
+			Context superParentContext = parentContext.getParent();
+			
 			if (parentContext.getContextClass().equals(StockMove.class)) {
 				stockMove = parentContext.asType(StockMove.class);
+			} else if(superParentContext.getContextClass().equals(StockMove.class)) {
+				stockMove = superParentContext.asType(StockMove.class);
 			} else {
 				return;
 			}
 		}
+		
 		stockMoveLine = stockMoveLineService.compute(stockMoveLine, stockMove);
 		response.setValue("unitPriceUntaxed", stockMoveLine.getUnitPriceUntaxed());
 		response.setValue("unitPriceTaxed", stockMoveLine.getUnitPriceTaxed());
@@ -73,4 +82,16 @@ public class StockMoveLineController {
         }
     }
 
+	
+	public void emptyLine(ActionRequest request, ActionResponse response){
+		StockMoveLine stockMoveLine = request.getContext().asType(StockMoveLine.class);
+		if(stockMoveLine.getLineTypeSelect() != StockMoveLineRepository.TYPE_NORMAL){
+			Map<String,Object> newStockMoveLine =  Mapper.toMap(new StockMoveLine());
+			newStockMoveLine.put("qty", BigDecimal.ZERO);
+			newStockMoveLine.put("id", stockMoveLine.getId());
+			newStockMoveLine.put("version", stockMoveLine.getVersion());
+			newStockMoveLine.put("lineTypeSelect", stockMoveLine.getLineTypeSelect());
+			response.setValues(newStockMoveLine);
+		}
+	}
 }