+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/SaleOrderStockServiceImpl.java
@@ -236,7 +236,11 @@ public class SaleOrderStockServiceImpl implements SaleOrderStockService  {
 	            stockMoveLine.setSaleOrderLine(saleOrderLine);
 	            stockMoveLine.setReservedQty(saleOrderLine.getReservedQty());
 			}
-
+			
+			stockMoveLine.setLineTypeSelect(saleOrderLine.getTypeSelect());
+			stockMoveLine.setPackPriceSelect(saleOrderLine.getPackPriceSelect());
+			stockMoveLine.setIsSubLine(saleOrderLine.getIsSubLine());
+			
 			return stockMoveLine;
 		}
 		else if(saleOrderLine.getTypeSelect() == SaleOrderLineRepository.TYPE_PACK){