+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/invoice/generator/InvoiceLineGeneratorSupplyChain.java
@@ -27,6 +27,7 @@ import com.axelor.apps.account.db.InvoiceLine;
 import com.axelor.apps.account.db.TaxLine;
 import com.axelor.apps.account.db.repo.AnalyticMoveLineMngtRepository;
 import com.axelor.apps.account.db.repo.AnalyticMoveLineRepository;
+import com.axelor.apps.account.db.repo.InvoiceLineRepository;
 import com.axelor.apps.account.service.invoice.InvoiceLineService;
 import com.axelor.apps.account.service.invoice.InvoiceToolService;
 import com.axelor.apps.account.service.invoice.generator.InvoiceLineGenerator;
@@ -35,7 +36,6 @@ import com.axelor.apps.base.db.Unit;
 import com.axelor.apps.base.service.UnitConversionService;
 import com.axelor.apps.purchase.db.PurchaseOrderLine;
 import com.axelor.apps.sale.db.SaleOrderLine;
-import com.axelor.apps.sale.db.repo.SaleOrderLineRepository;
 import com.axelor.apps.stock.db.StockMove;
 import com.axelor.apps.stock.db.StockMoveLine;
 import com.axelor.apps.supplychain.service.app.AppSupplychainService;
@@ -59,17 +59,17 @@ public abstract class InvoiceLineGeneratorSupplyChain extends InvoiceLineGenerat
 										   BigDecimal qty, Unit unit, TaxLine taxLine,
 										   int sequence, BigDecimal discountAmount, int discountTypeSelect,
 										   BigDecimal exTaxTotal, BigDecimal inTaxTotal, boolean isTaxInvoice,
-										   SaleOrderLine saleOrderLine, PurchaseOrderLine purchaseOrderLine, StockMoveLine stockMoveLine) {
-		super(invoice, product, productName, price, priceDiscounted, description, qty, unit, taxLine, sequence, discountAmount, discountTypeSelect, exTaxTotal, inTaxTotal, isTaxInvoice);
+										   SaleOrderLine saleOrderLine, PurchaseOrderLine purchaseOrderLine, StockMoveLine stockMoveLine, boolean isSubLine) {
+		super(invoice, product, productName, price, priceDiscounted, description, qty, unit, taxLine, sequence, discountAmount, discountTypeSelect, exTaxTotal, inTaxTotal, isTaxInvoice, isSubLine);
 		this.saleOrderLine = saleOrderLine;
 		this.purchaseOrderLine = purchaseOrderLine;
 		this.stockMoveLine = stockMoveLine;
 	}
 
 	protected InvoiceLineGeneratorSupplyChain(Invoice invoice, Product product, String productName, String description, BigDecimal qty,
-											  Unit unit, int sequence, boolean isTaxInvoice, SaleOrderLine saleOrderLine, PurchaseOrderLine purchaseOrderLine, StockMoveLine stockMoveLine) throws AxelorException {
+											  Unit unit, int sequence, boolean isTaxInvoice, SaleOrderLine saleOrderLine, PurchaseOrderLine purchaseOrderLine, StockMoveLine stockMoveLine, boolean isSubLine) throws AxelorException {
 
-		super(invoice, product, productName, description, qty, unit, sequence, isTaxInvoice);
+		super(invoice, product, productName, description, qty, unit, sequence, isTaxInvoice, isSubLine);
 
 		this.saleOrderLine = saleOrderLine;
 		this.purchaseOrderLine = purchaseOrderLine;
@@ -81,10 +81,13 @@ public abstract class InvoiceLineGeneratorSupplyChain extends InvoiceLineGenerat
 			this.priceDiscounted = saleOrderLine.getPriceDiscounted();
 			this.taxLine = saleOrderLine.getTaxLine();
 			this.discountTypeSelect = saleOrderLine.getDiscountTypeSelect();
-			this.isTitleLine = saleOrderLine.getTypeSelect() == SaleOrderLineRepository.TYPE_PACK;
-		} else if (purchaseOrderLine != null){
-			this.isTitleLine = purchaseOrderLine.getIsTitleLine();
-			this.purchaseOrderLine = purchaseOrderLine;
+			this.typeSelect = saleOrderLine.getTypeSelect();
+		} 
+		else if (purchaseOrderLine != null){
+		  if(purchaseOrderLine.getIsTitleLine()) {
+		      this.typeSelect = InvoiceLineRepository.TYPE_TITLE; 
+		  }			
+		  this.purchaseOrderLine = purchaseOrderLine;
 			this.discountAmount = purchaseOrderLine.getDiscountAmount();
 			this.price = purchaseOrderLine.getPrice();
 			this.priceDiscounted = purchaseOrderLine.getPriceDiscounted();
@@ -104,7 +107,7 @@ public abstract class InvoiceLineGeneratorSupplyChain extends InvoiceLineGenerat
 		}
 		
     }
-	
+
 
 	/**
 	 * @return