+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/web/InvoiceController.java
@@ -35,6 +35,7 @@ import com.axelor.apps.base.db.Wizard;
 import com.axelor.apps.sale.db.SaleOrder;
 import com.axelor.apps.supplychain.service.SaleOrderInvoiceService;
 import com.axelor.apps.supplychain.service.app.AppSupplychainService;
+import com.axelor.apps.supplychain.service.invoice.InvoiceServiceSupplychain;
 import com.axelor.apps.supplychain.service.invoice.SubscriptionInvoiceService;
 import com.axelor.db.JPA;
 import com.axelor.exception.service.TraceBackService;
@@ -58,6 +59,9 @@ public class InvoiceController {
 	
 	@Inject
 	protected SubscriptionInvoiceService subscriptionInvoiceService;
+
+	@Inject
+	protected InvoiceServiceSupplychain invoiceService;
 	
 	public void fillInLines(ActionRequest request, ActionResponse response) {
 		Invoice invoice = request.getContext().asType(Invoice.class);
@@ -282,7 +286,26 @@ public class InvoiceController {
 					e.getMessage()));
 			TraceBackService.trace(e);
 		}
-		
 	}
+		
 	
+	public void addSubLines(ActionRequest request, ActionResponse response) {
+  	try {
+          Invoice invoice = request.getContext().asType(Invoice.class);
+          response.setValue("invoiceLineList",  invoiceService.addSubLines(invoice.getInvoiceLineList()));
+      } catch (Exception e) {
+          TraceBackService.trace(response, e);
+          response.setReload(true);
+      }
+  }
+	
+	public void removeSubLines(ActionRequest request, ActionResponse response) {
+      try {
+          Invoice invoice = request.getContext().asType(Invoice.class);
+          response.setValue("invoiceLineList",  invoiceService.removeSubLines(invoice.getInvoiceLineList()));
+      } catch (Exception e) {
+          TraceBackService.trace(response, e);
+          response.setReload(true);
+      }
+  }
 }