+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/web/SaleOrderController.java
@@ -27,6 +27,8 @@ import java.util.List;
 import java.util.Map;
 
 import com.axelor.apps.account.db.Invoice;
+import com.axelor.apps.account.db.InvoiceLine;
+import com.axelor.apps.account.db.repo.InvoiceLineRepository;
 import com.axelor.apps.account.db.repo.InvoiceRepository;
 import com.axelor.apps.base.db.Company;
 import com.axelor.apps.base.db.Currency;
@@ -42,6 +44,8 @@ import com.axelor.apps.sale.db.repo.SaleOrderRepository;
 import com.axelor.apps.stock.db.StockLocation;
 import com.axelor.apps.stock.db.StockMove;
 import com.axelor.apps.stock.service.StockLocationService;
+import com.axelor.apps.stock.db.StockMoveLine;
+import com.axelor.apps.stock.db.repo.StockMoveLineRepository;
 import com.axelor.apps.supplychain.exception.IExceptionMessage;
 import com.axelor.apps.supplychain.service.SaleOrderCreateServiceSupplychainImpl;
 import com.axelor.apps.supplychain.service.SaleOrderInvoiceService;
@@ -65,6 +69,7 @@ import com.google.common.base.Joiner;
 import com.google.common.base.Strings;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
+import com.google.inject.persist.Transactional;
 
 @Singleton
 public class SaleOrderController{
@@ -89,6 +94,9 @@ public class SaleOrderController{
 						.createStocksMovesFromSaleOrder(saleOrderRepo.find(saleOrder.getId()));
 
 				if (stockMove != null) {
+				  
+	        this.generateStockMoveLineParentLine(stockMove.getStockMoveLineList());
+
 					response.setView(ActionView
 							.define(I18n.get("Stock move"))
 							.model(StockMove.class.getName())
@@ -102,6 +110,19 @@ public class SaleOrderController{
 			}
 		} catch (Exception e) {
 			TraceBackService.trace(response, e);
+		}
+	}
+	
+	@Transactional
+	public void generateStockMoveLineParentLine(List<StockMoveLine> stockMoveLines) {
+
+		for (StockMoveLine line : stockMoveLines) {
+			if (line.getSaleOrderLine() != null) {
+				line.setPackPriceSelect(line.getSaleOrderLine().getPackPriceSelect());
+				StockMoveLine parentStockMoveLine = Beans.get(StockMoveLineRepository.class).all()
+						.filter("self.saleOrderLine = ?", line.getSaleOrderLine().getParentLine()).fetchOne();
+				line.setParentLine(parentStockMoveLine);
+			}
 		}
 	}
 
@@ -235,6 +256,7 @@ public class SaleOrderController{
 					);
 
 			if(invoice != null)  {
+				this.generateInvoiceLineParentLine(invoice.getInvoiceLineList());
 				response.setCanClose(true);
 				response.setView(ActionView
 						.define(I18n.get("Invoice generated"))
@@ -250,6 +272,18 @@ public class SaleOrderController{
 		catch(Exception e)  { TraceBackService.trace(response, e); }
 	}
 	
+	@Transactional
+	public void generateInvoiceLineParentLine(List<InvoiceLine> invoiceLine) {
+		
+		for(InvoiceLine line : invoiceLine) {
+			if(line.getSaleOrderLine() != null) {
+				line.setPackPriceSelect(line.getSaleOrderLine().getPackPriceSelect());
+				line.setTotalPack(line.getSaleOrderLine().getTotalPack());
+				InvoiceLine parentInvoiceLine = Beans.get(InvoiceLineRepository.class).all().filter("self.saleOrderLine = ?", line.getSaleOrderLine().getParentLine()).fetchOne();
+				line.setParentLine(parentInvoiceLine);
+			}
+		}
+	}
 	
 
 	@SuppressWarnings({ "rawtypes", "unchecked" })